+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.26.4 (WIP)
+
+- Fixed `RecordErrorEvent.Error` and `CollectionErrorEvent.Error` sync with `ModelErrorEvent.Error` sync with ([#6639](https://github.com/pocketbase/pocketbase/issues/6639)).
+
+
 ## v0.26.3
 
 - Fixed and normalized logs error serialization across common types for more consistent logs error output ([#6631](https://github.com/pocketbase/pocketbase/issues/6631)).